Add a ModelFeature property test and fix old one-hot bug
The hypothesis test introduced in this commit caught a bug that has gone
unnoticed up until now; if the request-to-features code received a
timestamp that was exactly 10:00 PM, an IndexError would be raised. This
is more likely than it sounds, since requests occurring during the
10:00–10:30 PM grace period are always changed to have 10:00 PM
timestamps prior to being passed to the model.